requirements.txt
pandas
openpyxl
pyarrow
seaborn
matplotlib

fpdf2